Depreciate CpcPerformancePeriodValidation and update PcfPerformancePeriodValidation
